Move TypeTraits into sum-internal.h
