+++ cpp/src/arrow/compute/kernels/aggregate-test.cc
@@ -23,6 +23,7 @@
 #include "arrow/array.h"
 #include "arrow/compute/kernel.h"
 #include "arrow/compute/kernels/sum.h"
+#include "arrow/compute/kernels/sum-internal.h"
 #include "arrow/compute/test-util.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
